pyproject.toml
[project]
name = "sphinx_pieces"
version = "0.1.0"
requires-python = ">=3.9"
# Temporarily using version less than 8 just to test it.
dependencies = ["pyyaml>=6.0.2", "sphinx>=7.4.7,<8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-sphinx>=0.6.0",
    "types-docutils>=0.21.0.20250710",
]
